Begin moving syscall definitions into libc.
Allowing apps to directly invoke syscalls is a layering violation that
makes it difficult to port libc and do other cleanups. This CL moves
syscall declarations into libc. Apps still call the syscalls directly,
but this change sets up for renaming the syscalls and implementing
thunks in libc.

References to trusty_std.h also need to be eliminated, as the main use
for this file is directly referencing the syscalls.

Also opportunistically reorder includes.

Bug: 110159511

Change-Id: I7f0c47e727400d5c6a30ec38c534ba90bae638b3