Prefix syscalls with _trusty_* and add stubs in libc.
This change makes it easier to modify the parameter and return types of
these functions. It also makes it easier to port an off-the-shelf libc
to Trusty.

trusty_syscalls.h is updated to reflect changes in stubgen.py - not just
changing the stub names, but also forward declaring structs and including
stdint.h.

Bug: 110159511
Change-Id: Ifb35f37ba720fed8bb9e3518f697cd11100126ad