Add support and switch to use dynamically generated syscall stubs
Syscall stubs are  generated by stubgen tool and used to be manually
checked in into source tree. This CL adds /lib/syscall-stubs module
that does this dynamically at compile time which eliminate the need
for manual step. It also removes checked in versions syscall stubs
as thet are not need anymore.

Bug: 119623911
Change-Id: Iaad1f761858386d28c53549b7ab12f1e7881a00b