Unify definition of Trusty logging macros
This CL adds centralized definition of TLOGX macros that can be used
by applications to emit log messages:
   TLOGC - critical error messages, typically unrecoverable errors
   TLOGE - normal error messages, typically recoverable errors
   TLOGW - warning messages, typically unexpected non-fatal conditions
   TLOGI - other informational messages
   TLOGD - generic debug messages

The caller has to provide TLOG_TAG #define (typically app or a module
specific string) that would identify application or module emitting
messages. Log level is controlled by TLOG_LVL #define that could be
set to one of the following values:
   TLOG_LVL_NONE  - no messages are emitted
   TLOG_LVL_CRIT  - only critical error messages are emitted
   TLOG_LVL_ERROR - normal or critical error messages are emitted
   TLOG_LVL_WARN  - warnings or higher priority messages are emitted
   TLOG_LVL_INFO  - info or higher priority messages are emitted
   TLOG_LVL_DEBUG - debug or higher priority messages are emitted

If TLOG_LVL is not defined it will be set to TLOG_LVL_INFO.
The TLOG_LVL can be redefined by local module to facilitate local
debug.

Bug: 119891603
Change-Id: I0aa11a6436c87c3ad8dc0fabcf7be6cf0f1a4884