Build trusty-libc with -ffreestanding
We #define away the "dl" prefix on all the dlmalloc functions. This
results in a situation where to check whether it *needs* to zero all the
memory, calloc() will after calling malloc() subtract from pointer, as a
char*, in order to access the memory header to check its pre-zeroed
status.

Clang is rather aggressive in how it interprets the spec, and sees the
call to *malloc* rather than *dlmalloc*, and realizes it is with its
rights to treat reads that it knows are out of bounds (such as to the
memory header) as undefined. It thus arbitrarily decides that it got the
"fast" result (the memory was pre-zeroed) and deletes the codepath that
initializes the buffer.

By passing -ffreestanding, we instruct Clang to make fewer assumptions
about the behavior of functions called "malloc".

Bug: 126579822
Change-Id: Iae9f8d7c0262073be3dd68c6fa47554b2490ecbf