Clean up type signatures of syscalls.
Instead of returning long values, prefer int.

If a function returns a byte count or an error, return ssize_t.

Make handle_t signed so that it can be used as a return value with
possible error.

If an argument is a handle, make it a handle_t rather than uint32_t.

This makes the signature of close(...) compatible with POSIX.

Bug: 138738352
Change-Id: Iac2ec408c70e8fd37f857d834693e5aa3feb2c33