Move extra libc headers into a subdirectory
The augmented Trusty libc headers conflict with the musl headers with
the same name. We can't add both to the SDK sysroot without renaming or
moving them.

Copies time.h, memref.h, sys/mman.h, and sys/wait.h into the trusty/
folder under includes and updates references accordingly. This commit
does not delete the original headers so that we can make sure that all
references are updated and builds do not break during this process. When
all references are updated to the new paths we will remove the original
headers.

Test: build.py qemu-generic-arm64-test-debug
Bug: 169448829
Change-Id: I893a3922f7d59317ad471ed9edca68f6553a705f