Enable fake threading, enough to support mutexes
Some applications which will still work without the ability to create
threads assume the ability to link in mutex support and similar code. By
providing stubs for pthread calls and enabling the C++ features, we can
allow this code to run. Attempting to actually create an additional
thread is still impossible.

Bug: 152900132
Change-Id: Idb90975c3e4cbbe9a2583fccf48765cd5d93f993