Fix tipc test when built with coverage.
Building with coverage creates a connection to the coverage aggregator
service from each app during initialization. The tipc test hardcoded the
number of expected open ports, so fails when that coverage connection is
added. Fix the tipc unittests to not assume that it is the only module
opening IPC handles.

Test: build.py qemu-generic-arm64-fuzz-test-debug
Bug: 179931565
Change-Id: I827db110b18b0d424f0d3834d5cac0c864556e28