Increase stack and heap for ACVP test app
Boringssl's ECDH routines require more stack and heap space when used in
ACVP tests. Experimentally found that 16M for each is sufficient, but 8M
is not. Unfortunately the boringssl modulewrapper for ACVP does not
handle allocation failures gracefully. This is acceptable as long as the
ACVP app is only used in test builds.

Test: adb shell acvptool -wrapper trusty_acvp_modulewrapper -json vectors/KAS-ECC-SSC
Bug: None
Change-Id: Ic2ad189ce57e3215b3600901135eb26822bc8e0a