Handle logging for unittests without HAS_ define
The legacy build system used USER_TASK_WITH_ defines to check if the
unittest library was included in an app, and if so called
unittest_printf for every TLOG. This is not compatible with the new,
modularized build userspace build system, since we don't necessarily
know when building a module whether the app will include the unittest
library.

Replace the unittest_printf function with a _tlog function that is
defined in both libunittest and in libc. When libunittest is earlier
than libc in the link command its version of the function will be used,
otherwise the libc version that just prints to standard error will be
used. All libraries which depend on libc, including libunittest, will
always be earlier in the link order because we need to find undefined
symbol uses before scanning the static library that defines those
symbols.

Test: build.py qemu-generic-arm64-test-debug
Bug: 183754680
Change-Id: I86ff705a6e2734af1b3d1b677b27074cb9956113