[app][apploader] Fix BN_bn2bin return value checks
Because we have to be compatible with OpenSSL 1.0.2, the signature
code uses BN_bn2bin instead of BN_bn2binpad. The former returns
the size of the actual big-endian number without zero padding,
which may be smaller than the full size of the buffer for
ECDSA signatures. This changes the checks on the return value of
BN_bn2bin to check against the actual size, which fixes sporadic
failures when signing packages.

Bug: 115420908
Change-Id: I6817da43203bdc177b1754664699da0ef95cd104