Generate constant headers for libraries
The manifest compiler builds both constant headers and the manifest
binary. Some libraries contain constant headers but libraries don't have
a manifest file, which is only for apps. Since we want libraries to be
standalone, we need to generate the constants header for a library
before building that library rather than waiting until it is included in
an app.

Test: build.py qemu-generic-arm64
Bug: 169448829
Change-Id: I9da866f7fb2bd23dc43c99dace3c43e58c47cb97