Fix recursive dependency libraries list
Recursive dependencies were not correctly picking up their dependency
libraries. For the dependency chain A -> B -> A, library B would not
include a dependency on A's static library. Fix this by adding A's
library to its cached variables early, so that B picks up the library
while recursing into its dependencies in userspace_recurse.mk.

This happened to work because recursive dependencies were previously
always added together, so we didn't notice this issue. While adding Rust
libraries, we found that libc's dependencies were incorrect.

Removes EXPORT_OBJECTS, which was a confusing, unused variable.

Fixes dependency's extra objects to always be propagated up. libc
exports an extra object for crti, which needs to be propagated up to the
linked app, and this was only happening if libc was a direct dependency.

Test: Add a dependency on libc but not ubsan to a module, build
Bug: 169448829
Change-Id: Ibdc6f6825ed9b613d6fcdb9b3861f5172951e923