[acvp] Fix output length and fixed data for KDF
The supportedLengths parameter for the KDF algo is in bits, not bytes,
so we need to specify that we only accept byte sized output lengths.

The ACVP server expects the L parameter (output length in bits) of the
PRF input to also be included in the fixed data. Add L to the fixed data
returned as part of the answer.

The ACVP demo server KDF test now passes.

Test: build.py
Test: acvptool -wrapper trusty_acvp_modulewrapper -run KDF
Bug: 193737836
Change-Id: Ifb8073ad2d5fed6b280b7d34ab340745b81545d3