[rng] Replace trusty_rng_secure_rand with the BoringSSL RNG
Replaces the Trusty RNG library with the BoringSSL RNG by forwarding the
existing API to BoringSSL. The add_entropy() API does not add entropy
directly to the BoringSSL RNG entropy pool, but instead adds entropy to
an internal system RNG that BoringSSL uses as a source of additional
randomness.

Concretely, this change forwards the trusty_rng_secure_rand API directly
to BoringSSL and exposes the prior functionality as an internal API just
for BoringSSL to provide fast system randomness.

Test: build.py
Bug: 193674299
Change-Id: I82fabe2e0ad7c3b5cbc92354fb4f436f2afbda0c