[rust] Add test framework service
Adds a tipc service compatible with the existing test clients to the
test crate. When a client connects to the designated port for a crate,
all the tests for that crate will run.

Currently tests can't fail, as we have no way to track failures other
than aborting the test service. This will need global state to track
failures and custom macros for asserts, as in the existing C unittest
framework.

Test: build.py
Bug: 195994856
Change-Id: I8b0c29d58c115908eea3e4dc1af08e274f7ae4e5