[unittest-rust] Add assert macros for use in Rust tests
Rust tests usually fail by panicking, which is caught and noted by the
test runtime. We aren't using panic=unwind in Trusty, so we can't
recover from panics. Instead we want to use test failure macros that
record the failure in global state for the test harness to later report.

Adds assert! and expect! macros that behave like the ASSERT and EXPECT
macros from the C test framework.

Test: build.py
Bug: 195994856
Change-Id: I3bcab871fd8f289bf7583984080edc3d7dac7aa6