Fix exported flags for leaf modules
Exported flags are intended to always be included in the local modules
flags as well. However, exported flags were previously appended in
userspace_recurse.mk, and therefore would only be added to the local
build if the module had dependencies. Instead, we now append exported
flags to the local build in library.mk after all dependencies have been
resolved. This is equivalent to the previous behavior if the module had
any dependencies, but now the behavior is consistent across modules with
and without dependencies.

Test: build.py
Bug: None
Change-Id: Icc6da8bf543cce0c51b072eecfef1f3a47839235