[tools][manifest_compiler] Fix UUID tests
Rename second instance of the test function test_validate_uuid_7 to
test_validate_uuid_8. The first instance contained a broken test so
it was repurposed to increase test coverage of the parse_uuid function
in the manifest compiler. Error messages reworded and updated to use
formatting strings for improved readability.

Bug: 197031257
Change-Id: I503234096a8f698a4b5616fd2207a677f63d4346