Fix dependencies for syscall stubs and headers
There were multiple targets for the syscall stubs rules, but this
resulted in multiple calls to the stub generator. This created a race
condition between a second call to the stub generator and uses of the
generated header, and if the use of the header happened to read the file
while the python generator script was writing to the file, it could see
an empty or partial file.

Makes the stubgen make rule have a single target with other generated
files depending on that target.

Test: build.py
Bug: None
Change-Id: Ied8f89a3115fc98574922698c1e69156b91671b1