apploader: Replace cppbor with open-dice CBOR code
This change replaces all uses of cppbor, a C++ library with a high
level interface to read and write CBOR data with a thin C++ wrapper
around open-dice. The result is that we make few if any temporary
allocations as we write out CBOR data which should allow operation
in memory-constrained environments.

Test: build.py qemu-generic-arm64-test-debug --test apploader \
    --test appmngr
Bug: 2276522071
Change-Id: I866b4cd3ce5aa8cb854d7e90c5eb0984bb8c862e