[rust] Use MODULE_RUST_EDITION to override the default Rust edition
Rather than having to explictly add the --edition flag for crates which
do not use the default edition, we now support a MODULE_RUST_EDITION
variable which will cause the module to be built with the specified
edition. If unspecified, it is assumed to be the default edition
configured in common_flags.mk.

Test: build.py
Bug: 192309116
Change-Id: Ib346022c9d80f0b953fbb0b8bbe2ede54efff7e6