[crash-test] Fix nx data test after LLVM semantic changes
LLVM used to not care that we were putting a function in purely data
sections (.data and .rodata), and would happily place the executable
instructions in a non-executable well-known data section.
https://reviews.llvm.org/D100944 changed this behavior - explicit
section assignment to a section with conflicting flags (in our case,
sections without SHF_EXECINSTR) results in the function being assigned
to a new, unique section with the executable flag. The linker then
merges permissions for the unique executable .data and .rodata sections
and the normal data-only versions of those sections, and sets the entire
section to be executable.

To work around this, this change moves the data functions into their own
module built without LTO, then rewrites the section flags on .data and
.rodata to revert them back to non-executable.

Test: com.android.trusty.crashtest
Bug: 185914685
Change-Id: I71e7f6ec37afb52c0b95aa46784dc67eed85339b