[rust][tipc] Do not drop channel cookie until connection is closed
The Weak channel cookie stored in a handle set context is owned by the
handle set, and should not be dropped when accessed during handle
operations. This change fixes the lifetime of this pointer so that it
lives as long as the channel is active and attached to a handle set.

Test: build.py --test com.android.trusty.rust.tipc
Bug: 230823197
Change-Id: I35806c3fb938ea000bf2074cf5880f9073ccd96f