tools: Add apploader flags to manifest compiler
Add a new flag group to the manifest compiler to control loading
of apps. Add a new flag which can be used to ensure that the ELF
image of an app is encrypted before it is loaded. Add a new test
case to ensure we can successfully roundtrip between the JSON
and binary representations of the new flag group.

Bug: 241824652
Test: python3 -m unittest -v test_manifest_compiler
Change-Id: I5bcdc8099d1b6c094044c1ee18e4317625cd4a69