app/apploader: Enforce app encryption requirement
Read the requires_encryption manifest property and do not load an
application if it requires encryption but the ELF item was not
a COSE_Encrypt structure. The apploader_parse_package_metadata
function already detects whether the application package is using
encryption so we simply extend the apploader_package_metadata
struct to record whether the package was decrypted during parsing.

This change also consolidates the manifest parsing code in its own
translation unit: app_manifest.

Bug: 241824652
Change-Id: I56103d528cc02498af95b5e1ee211e6ad9eed233