apploader: Always print "encryption required" status in package tool
Currently, the "encryption required" flag is only inspected if the app is not encrypted; this catches egregious mistakes. If the app is encrypted, there is no easy way to determine if the "encryption required flag" was enabled in the manifest. Since an encrypted app without the "encryption required" flag may be one build change away from disaster, we would like the host tool to always show whether "encryption required" has been enabled.

Output is: "Encryption: {YES|NO}, {OPTIONAL|REQUIRED}"

Test: apploader_package_tool --mode info g7.app
Bug: 241824652
Change-Id: I08e57021151e36c47333f7b420f0eeaa9a65bdef