storage: Add STORAGE_FILE_GET_MAX_SIZE API on Trusty side
The storage TA hardcodes that backing store files can become as
large as 0x10000000000 bytes. This works when there is sparse
file support. As part of moving backing stores to our own
block device, we want to plumb the max size from Android
instead. This CL defines the API; a corresponding API CL
(and implementation) will be needed on the Android side.

Bug: 247013568
Test: None
Change-Id: I7da109b2e43e10558d17327735d1208cbf7056f6