lib/libc-trusty: Modify libc tests for the kernel
This modifies the userspace libc tests to allow reusing them
in the kernel and defines a new port test
"com.android.kernel.libctest". While most tests that are
 #ifdef'ed out in this CL are not applicable to the kernel
(e.g. printing floating-point, sbrk, using types like time_t
which are not defined by musl in the kernel) some of the
removed tests may be able to run in the kernel and should be
added back eventually (e.g. shadow_call_stack).

Bug: 230134581
Change-Id: If554112c8b34837ed9036e3a5309316ee16d3716