tools: Adjust how mypy is invoked as a pre-upload script
Enable pre-upload hook to run type checks on committed files,
running mypy one file at a time instead of for the whole project.

If mypy is run with a directory argument, it is assumed that all
files in that directory and its subdirectories are part of the
same module [0]. This causes errors if a python file with the same
name exists in multiple folders. We instead pass the list of
python scripts that exist in the tree. This sidesteps the issue.

Since script repo hooks can only be a single command, we call
use `sh -c` to run a sequence of commands. Because a commit
can delete files, we filter out any files that do not exist
in the source tree before invoking mypy.

[0] https://mypy.readthedocs.io/en/stable/running_mypy.html#mapping-file-paths-to-modules

Bug: 240356613
Change-Id: Ifc9394f7062d1784f54aea547ff0d77808e98c04