lib/tipc/test: Improve tipc echo service
Several tipc tests send messages to an echo service. The echo
service used to interleave read and send operations which raises
warnings in the kernel ipc code. To avoid these warnings, we now
first read all messages into a buffer on the stack and then send
them. Before, the service would return when no more messages can
be sent since the message queue is full. Now we wait on the peer
to drain the message queue before resending messages. This means
that the service handler does not retain state between two calls.

In order to be able to store all messages before we send them
back to the peer, we increase the minimumi heap size setting
in the manifest.

Bug: 258851590
Test: build.py qemu-generic-arm64-test-debug --test echo-long --verbose
Change-Id: I10b81b3be43c04912ae4e85a98cadd95cf71ac3f