app/device_tree/tests: Make the device tree test blob an overlay
This commit adds an empty base device tree blob and changes
the test blob into an overlay. When testing on QEMU the
overlay is first applied to the empty base blob, then the
tests search for the node with compatible string
"google,test_node". The tests required a few small changes
since the order of nodes and properties in the overlay is
not preserved when the overlay is applied.

Bug: 231151995
Change-Id: I04f93e781874cffc77a3738ef90136b2865533ab