Correctly format config descriptor in degenerate chain
The RKP HAL specifies that degenerate chains can omit the config
descriptor, but if it's present then it should conform to the specified
map definition. Replace the non-compliant config descriptor consisting
of 64 bytes of zeros with a compliant descriptor.

The descriptor ends up being an empty map, i.e. a single 0xA0 byte, but
the utility functions are used in case we want to do something more in
the future and as an example of how to use them.

Bug: 266960035
Change-Id: I5daa11b793a0f543bc35d708d442fcccb55b1e7b