bssl-sys: Remove functions that use va_list
__builtin_va_list on aarch64 is defined as [u64; 4], but passing an
array "by value" is not allowed in Rust FFI. We don't need these
functions, so just drop them from the bindings.

Note: in Android we use --cap-lints allow for external/ so we don't see
these warnings. We could do that here, but better to ensure we don't
inadvertently break ABI here.

Test: build.py generic-arm64-test-debug
Bug: 271248782
Change-Id: Iccf62fbaf7d47cf01cee631599419e2aef13c1cd