make/library.mk: Add MODULE_RUST_STEM
Add a new variable that can be used to change the name
of the output rlib for a Rust crate. The file is called
lib$(MODULE_RUST_STEM).rlib if the variable is not empty,
otherwise it defaults to the crate name.

Append the "-C metadata" compiler flag to distinguish between
multiple crates with the same name but different stems.

Bug: 242243245
Test: build.py
Change-Id: I808749e10a5b26b46898a185969cb44224cd0cea