apploader: Indicate that the apploader service is enabled
Building all tests as loadable apps by default does not work if the
target build does not support loadable apps. We should not build any
loadable apps for build configurations that do not include the apploader
service, as there is no way to run these apps. This change enables the
build system to determine if the apploader service is included, without
having to depend on it existing in a particular path.

Test: build.py
Bug: 296257715
Change-Id: Idc146284b8c0e16894fcedc7295d66e297ecc76b